purchase regions
