+++ src/chopsticks.init.ts
@@ -1,5 +1,5 @@
 import { ApiPromise, WsProvider, Keyring } from "@polkadot/api";
-import { KeyringPair } from "@polkadot/keyring/types";
+import { purchaseRegion } from "./common";
 import { cryptoWaitReady } from "@polkadot/util-crypto";
 import { Timeslice } from "./types";
 import * as consts from "./consts";
@@ -33,23 +33,6 @@ async function startBulkSale(rococoApi: ApiPromise, coretimeApi: ApiPromise) {
   // Alice buys a region.
   await purchaseRegion(coretimeApi, alice);
   console.log("Region purchased");
-}
-
-async function purchaseRegion(
-  coretimeApi: ApiPromise,
-  buyer: KeyringPair,
-): Promise<void> {
-  const callTx = async (resolve: () => void) => {
-    const purchase = coretimeApi.tx.broker.purchase(consts.INITIAL_PRICE * 2);
-    const unsub = await purchase.signAndSend(buyer, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
 }
 
 async function setStatus(
